+++ syncopy/datatype/base_data.py
@@ -991,11 +991,9 @@ class BaseData(ABC):
     @property
     def trialdefinition(self):
         """nTrials x >=3 :class:`numpy.ndarray` of [start, end, offset, trialinfo[:]]"""
-        return np.array(self._trialdefinition)
-
-    @trialdefinition.setter
-    def trialdefinition(self, trl):
-        _definetrial(self, trialdefinition=trl)
+        if self._trialdefinition is not None:
+            # to avoid hanging references
+            return self._trialdefinition.copy()
 
     @property
     def sampleinfo(self):
@@ -1010,6 +1008,12 @@ class BaseData(ABC):
         raise SPYError(
             "Cannot set sampleinfo. Use `BaseData.trialdefinition` instead."
         )
+
+    @property
+    def trial_ids(self):
+        """Index list of trials"""
+        if self._trialdefinition is not None:
+            return self._trial_ids
 
     @property
     def trialintervals(self):
@@ -1034,11 +1038,11 @@ class BaseData(ABC):
 
     @property
     def trials(self):
-        """list-like array of trials"""
+        """list-like iterable of trials"""
 
         if self.sampleinfo is not None:
             trial_ids = list(range(self.sampleinfo.shape[0]))
-            # this is cheap as it just initializes a list-like object
+            # this is cheap as it just initializes an indexable generator
             # with no real data and/or computation!
             return TrialIndexer(self, trial_ids)
         else: